chore(rust): remove isolating finalizer
