+++ crates/rolldown/src/module_finalizers/hmr.rs
