+++ crates/rolldown/src/module_finalizers/impl_visit_mut.rs
