+++ crates/rolldown_common/src/inner_bundler_options/types/output_format.rs
@@ -18,10 +18,6 @@ pub enum OutputFormat {
 }
 
 impl OutputFormat {
-  pub fn requires_scope_hoisting(&self) -> bool {
-    matches!(self, Self::Esm | Self::Cjs | Self::Iife | Self::Umd)
-  }
-
   #[inline]
   pub fn keep_esm_import_export_syntax(&self) -> bool {
     matches!(self, Self::Esm)
